requirements.txt
transformers
git+https://github.com/openai/whisper.git
torch
accelerate
gradio
Pillow
matplotlib
bitsandbytes
gtts
torchvision
